Add tip padding to lower link and rename tip link
Add the mesh of the tip padding to the lower link instead of the padding
link.  This way, there is no collision between the meshes of the two
links.

Keep the tip link (which does not have any geometry anymore) as it is
useful to get the tip position.  However, rename it to
`finger_tip_link`, i.e. remove the "padding" from the name.