Set different colours for the fingers
Add a "material" option to creating visuals and fingers.  This way the
three fingers of the TriFinger can be assigned different colours (0:
red, 120: green, 240: blue) to easily distinguish them.

Unfortunately I did not find a way to have a default material (defining
it in the macros file will lead to name collisions when including it
multiple times...), so this is a mandatory argument now.