Convert to Python package
Convert this package from CMake-based build to pure Python.  For this,
the xacro files are now build using a custom build command in setup.py

This change will make it possible to publish robot_properties_fingers on
PyPI, so that trifinger_simulation can depend on it and thus won't need
to mirror the model files anymore.